git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/trunk@1403782 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/core/src/java/org/apache/lucene/search/BooleanScorer2.java
@@ -129,7 +129,7 @@ class BooleanScorer2 extends Scorer {
     }
 
     @Override
-    public float freq() throws IOException {
+    public int freq() throws IOException {
       return 1;
     }
 
@@ -313,7 +313,7 @@ class BooleanScorer2 extends Scorer {
   }
 
   @Override
-  public float freq() throws IOException {
+  public int freq() throws IOException {
     return countingSumScorer.freq();
   }
