+++ lucene/core/src/java/org/apache/lucene/search/ScoreCachingWrappingScorer.java
@@ -61,7 +61,7 @@ public class ScoreCachingWrappingScorer extends Scorer {
   }
 
   @Override
-  public float freq() throws IOException {
+  public int freq() throws IOException {
     return scorer.freq();
   }
