+++ lucene/queries/src/java/org/apache/lucene/queries/function/ValueSourceScorer.java
@@ -88,7 +88,7 @@ public class ValueSourceScorer extends Scorer {
   }
 
   @Override
-  public float freq() throws IOException {
+  public int freq() throws IOException {
     return 1;
   }
 }
